backend/pyproject.toml
[tool.poetry]
name = "ai-language-tutor-backend"
version = "0.1.0"
description = "AI Language Tutor Backend API - FastAPI application with multi-provider AI integration"
authors = ["AI Language Tutor Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
neo4j = "^5.15.0"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.4"
alembic = "^1.13.0"
openai = "^1.3.0"
google-genai = "^0.8.0"
google-cloud-speech = "^2.21.0"
google-cloud-texttospeech = "^2.16.3"
ginza = "^5.1.2"
spacy = "^3.7.2"
fugashi = "^1.5.2"
unidic = "^1.1.0"
httpx = "^0.25.2"
structlog = "^23.2.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.35.0"}
# Pin to a widely available version to satisfy Poetry resolver in CI/containers
reportlab = "4.2.0"
pykakasi = "2.2.1"
pandas = "^2.0.0"  # Needed for dictionary import

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
mypy = "^1.7.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
implicit_reexport = false
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "neo4j.*",
    "sqlalchemy.*",
    "psycopg2.*",
    "pgvector.*",
    "ginza.*",
    "spacy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    # Default to unit/contract tests; integration tests require extra services/data.
    "-m",
    "not integration",
    "--cov=app",
    "--cov-report=term-missing:skip-covered",
    # Reason: repo-mounted `htmlcov/` may be owned by root from prior container runs.
    # Writing to /tmp avoids permission issues while keeping HTML output available.
    "--cov-report=html:/tmp/htmlcov",
    # Reason: repo-mounted `coverage.xml` may be owned by root from prior container runs.
    "--cov-report=xml:/tmp/coverage.xml",
    # NOTE: The codebase is large and currently includes many integration-style tests.
    # Keep a realistic baseline so the default suite can go green; raise over time.
    "--cov-fail-under=35",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"